+++ src/cartridge/mbc3.rs
@@ -93,7 +93,7 @@ pub mod mbc3 {
                 latched: 0,
             };
         }
-        fn rom_read(&mut self, addr: u16) -> u8 {
+        fn rom_read(&mut self, addr: usize) -> u8 {
             match addr {
                 0..=0x3FFF => self.rom[0][addr as usize],
                 0x4000..=0x7fff => self.rom[self.rom_bank_num][addr as usize],
@@ -101,7 +101,7 @@ pub mod mbc3 {
             }
         }
 
-        fn rom_write(&mut self, addr: u16, val: u8) {
+        fn rom_write(&mut self, addr: usize, val: u8) {
             match addr {
                 0..=0x1fff => self.ram_enable = val == 0xA,
                 0x2000..=0x3fff => {
@@ -133,7 +133,7 @@ pub mod mbc3 {
                 _ => unreachable!(),
             }
         }
-        fn ram_read(&mut self, addr: u16) -> u8 {
+        fn ram_read(&mut self, addr: usize) -> u8 {
             //So this is the space A000->BFFF. We only go away from this if we turn on the bank mode.
             if !self.ram_enable {
                 return 0xFF;
@@ -151,7 +151,7 @@ pub mod mbc3 {
                 return self.ram[self.ram_bank_num][addr as usize];
             }
         }
-        fn ram_write(&mut self, addr: u16, val: u8) {
+        fn ram_write(&mut self, addr: usize, val: u8) {
             if self.ram_enable {
                 if self.clock_mode > 8 {
                     let clock_result = self.clock.lock();
