+++ src/cartridge/mbc6.rs
@@ -16,17 +16,17 @@ pub mod mbc6 { //net de get
                 ram: [0; 0x2000],
             };
         }
-        fn rom_read(&mut self, addr: u16) -> u8 {
+        fn rom_read(&mut self, addr: usize) -> u8 {
             return self.rom[addr as usize];
         }
 
-        fn rom_write(&mut self, _addr: u16, _val: u8) {
+        fn rom_write(&mut self, _addr: usize, _val: u8) {
             ()
         }
-        fn ram_read(&mut self, addr: u16) -> u8 {
+        fn ram_read(&mut self, addr: usize) -> u8 {
             return self.ram[addr as usize];
         }
-        fn ram_write(&mut self, addr: u16, val: u8) {
+        fn ram_write(&mut self, addr: usize, val: u8) {
             self.ram[addr as usize] = val;
         }
     }
