+++ src/cartridge/mbc7.rs
@@ -12,17 +12,17 @@ pub mod mbc7 { // accelerometer
             let _ = cart;
             todo!()
         }
-        fn rom_read(&mut self, addr: u16) -> u8 {
+        fn rom_read(&mut self, addr: usize) -> u8 {
             return self.rom[addr as usize];
         }
 
-        fn rom_write(&mut self, _addr: u16,_vall: u8) {
+        fn rom_write(&mut self, _addr: usize,_vall: u8) {
             ()
         }
-        fn ram_read(&mut self, addr: u16) -> u8 {
+        fn ram_read(&mut self, addr: usize) -> u8 {
             return self.ram[addr as usize];
         }
-        fn ram_write(&mut self, addr: u16, val: u8) {
+        fn ram_write(&mut self, addr: usize, val: u8) {
             self.ram[addr as usize] = val;
         }
     }
