+++ src/cartridge/mod.rs
@@ -67,8 +67,7 @@ pub mod cartridge {
         }   
     }
     impl AsMemory for Cartridge {
-        fn memory_map(&mut self, addr: u16) -> u8 {
-            //dbg!(addr);
+        fn memory_map(&mut self, addr: usize) -> u8 {
             if addr >= 0xA000 {
                 return match self {
                     Cartridge::Mbc0(mem) => mem.ram_read(addr - 0xA000),
@@ -94,7 +93,7 @@ pub mod cartridge {
             }
         }
 
-        fn memory_write(&mut self, addr: u16, val: u8) {
+        fn memory_write(&mut self, addr: usize, val: u8) {
             if addr > 0xA000 {
                 return match self {
                     Cartridge::Mbc0(mem) => mem.ram_write(addr - 0xA000, val),
